Update test_hydro.py (#11)
I added some tests, for example to test whether massbalance closes at each time step and not only at the end.
Other tests are more difficult to provide for different glaciers as glacier behaviour differed. I ran the test for some other glaciers in the Alps than HEF and some tests failed (I commented these out and provided a comment).
I added some test to check whether the overall behaviour seems correct when glacier area is declining. For the monthly run I added a test to see whether melt and liquid precipitation is larger in summer months than winter months. This is specific to European Alps. 

Do you normally run the test in OGGM only with HEF and assume that then it is correct for other glaciers as well or do you run the tests for a set of different glaciers worldwide?