added tests for `run_with_hydro_daily` and a notebook showing the results.
The TIModel can now run with the `run_with_hydro_daily` function from OGGM (https://github.com/OGGM/oggm/pull/1270).

I had to change couple of lines in `mbmod_daily_oneflowline` as well.

It works fine mostly.
However, Farinotti's volumes estimate are not matched in 2004, if the run starts in 1980. Not sure what the issue is.