add new function to process isimip data without correction. (#31)
This function might be a bit faster than other function because it does not load all the historical GCM data.
Also, if you have not prprocessed the historical GCM data for the glaciers, you can still process the future data with this function. (I haven't flattened my downscaled data for the past, because I don't expect to use bias correction).